requirements.txt
Flask
matplotlib
google-cloud-texttospeech
google-cloud
groq
google-generativeai
google-cloud-translate
gradio
pinecone
langchain-google-genai
rapidfuzz
langchain
pinecone
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
langchain
google-generativeai
langchain-google-genai